+++ testing/toxic/chat.c
@@ -88,7 +88,7 @@ static void chat_onKey(ToxWindow* self, int key) {
 
     wprintw(ctx->history, "%s\n", ctx->line);
 
-    if(m_sendmessage(ctx->friendnum, (uint8_t*) ctx->line, strlen(ctx->line)+1) < 0) {
+    if(m_sendmessage(ctx->friendnum, (uint8_t*) ctx->line, (uint32_t)strlen(ctx->line)+1) < 0) {
       wattron(ctx->history, COLOR_PAIR(3));
       wprintw(ctx->history, " * Failed to send message.\n");
       wattroff(ctx->history, COLOR_PAIR(3));
